improved structure, introduced a cron job to run in parallel and save snapHodlBalanceData
